Build and RPM GeoWave components using Docker containers

+++ deploy/packaging/rpm/admin-scripts/jenkins-build-geowave.sh
@@ -2,7 +2,12 @@
 #
 # GeoWave Jenkins Build Script
 #
-# In the Execute Shell block before calling this script set the versions
+
+echo "---------------------------------------------------------------"
+echo "         Building GeoWave with the following settings"
+echo "---------------------------------------------------------------"
+echo "BUILD_ARGS=${BUILD_ARGS} ${@}"
+echo "---------------------------------------------------------------"
 
 cd $WORKSPACE/deploy
 
@@ -44,41 +49,51 @@ $WORKSPACE/.utility/maven-jace-hack.sh
 
 # Build the debug bindings
 cd $WORKSPACE/deploy
-mvn package -P generate-jace-proxies,linux-amd64-gcc-debug $BUILD_ARGS "$@"
-mv $WORKSPACE/deploy/target/*-jace.jar $WORKSPACE/deploy/target/jace/geowave-jace.jar
-tar -czf $WORKSPACE/deploy/target/jace-linux-amd64-debug.tar.gz \
-    -C $WORKSPACE/deploy/target/jace geowave-ingest.jar \
-    -C $WORKSPACE/deploy/target/jace geowave-jace.jar \
-    -C $WORKSPACE/deploy/target/dependency/jace libjace.so \
-    -C $WORKSPACE/deploy/target/dependency/jace include
+if [ ! -f $WORKSPACE/deploy/target/jace-linux-amd64-debug.tar.gz ]; then
+    mvn package -P generate-jace-proxies,linux-amd64-gcc-debug $BUILD_ARGS "$@"
+    mv $WORKSPACE/deploy/target/*-jace.jar $WORKSPACE/deploy/target/jace/geowave-jace.jar
+    tar -czf $WORKSPACE/deploy/target/jace-linux-amd64-debug.tar.gz \
+        -C $WORKSPACE/deploy/target/jace geowave-ingest.jar \
+        -C $WORKSPACE/deploy/target/jace geowave-jace.jar \
+        -C $WORKSPACE/deploy/target/dependency/jace libjace.so \
+        -C $WORKSPACE/deploy/target/dependency/jace include
+fi
 
 # Build the release bindings
-mvn package -P generate-jace-proxies,linux-amd64-gcc-release $BUILD_ARGS "$@"
-tar -czf $WORKSPACE/deploy/target/jace-linux-amd64-release.tar.gz \
-    -C $WORKSPACE/deploy/target/jace geowave-ingest.jar \
-    -C $WORKSPACE/deploy/target/jace geowave-jace.jar \
-    -C $WORKSPACE/deploy/target/dependency/jace libjace.so \
-    -C $WORKSPACE/deploy/target/dependency/jace include
+if [ ! -f $WORKSPACE/deploy/target/jace-linux-amd64-release.tar.gz ] || [ ! -f $WORKSPACE/deploy/target/jace-source.tar.gz ]; then
+    mvn package -P generate-jace-proxies,linux-amd64-gcc-release $BUILD_ARGS "$@"
+    tar -czf $WORKSPACE/deploy/target/jace-linux-amd64-release.tar.gz \
+        -C $WORKSPACE/deploy/target/jace geowave-ingest.jar \
+        -C $WORKSPACE/deploy/target/jace geowave-jace.jar \
+        -C $WORKSPACE/deploy/target/dependency/jace libjace.so \
+        -C $WORKSPACE/deploy/target/dependency/jace include
 
-tar -czf $WORKSPACE/deploy/target/jace-source.tar.gz \
-    -C $WORKSPACE/deploy/target/jace geowave-ingest.jar \
-    -C $WORKSPACE/deploy/target/jace geowave-jace.jar \
-    -C $WORKSPACE/deploy/target/dependency/jace CMakeLists.txt \
-    -C $WORKSPACE/deploy/target/dependency/jace source \
-    -C $WORKSPACE/deploy/target/dependency/jace include
+    tar -czf $WORKSPACE/deploy/target/jace-source.tar.gz \
+        -C $WORKSPACE/deploy/target/jace geowave-ingest.jar \
+        -C $WORKSPACE/deploy/target/jace geowave-jace.jar \
+        -C $WORKSPACE/deploy/target/dependency/jace CMakeLists.txt \
+        -C $WORKSPACE/deploy/target/dependency/jace source \
+        -C $WORKSPACE/deploy/target/dependency/jace include
+fi
 
 # Build and archive HTML/PDF docs
 cd $WORKSPACE/
-mvn javadoc:aggregate
-mvn -P docs -pl docs install
-tar -czf $WORKSPACE/target/site.tar.gz -C $WORKSPACE/target site
+if [ ! -f $WORKSPACE/target/site.tar.gz ]; then
+    mvn javadoc:aggregate
+    mvn -P docs -pl docs install
+    tar -czf $WORKSPACE/target/site.tar.gz -C $WORKSPACE/target site
+fi
 
 # Build and archive the man pages
-mkdir -p $WORKSPACE/docs/target/{asciidoc,manpages}
-cp -fR $WORKSPACE/docs/content/manpages/* $WORKSPACE/docs/target/asciidoc
-find $WORKSPACE/docs/target/asciidoc/ -name "*.txt" -exec sed -i "s|//:||" {} \;
-find $WORKSPACE/docs/target/asciidoc/ -name "*.txt" -exec a2x -d manpage -f manpage {} -D $WORKSPACE/docs/target/manpages \;
-tar -czf $WORKSPACE/docs/target/manpages.tar.gz -C $WORKSPACE/docs/target/manpages/ .
+if [ ! -f $WORKSPACE/docs/target/manpages.tar.gz ]; then
+    mkdir -p $WORKSPACE/docs/target/{asciidoc,manpages}
+    cp -fR $WORKSPACE/docs/content/manpages/* $WORKSPACE/docs/target/asciidoc
+    find $WORKSPACE/docs/target/asciidoc/ -name "*.txt" -exec sed -i "s|//:||" {} \;
+    find $WORKSPACE/docs/target/asciidoc/ -name "*.txt" -exec a2x -d manpage -f manpage {} -D $WORKSPACE/docs/target/manpages \;
+    tar -czf $WORKSPACE/docs/target/manpages.tar.gz -C $WORKSPACE/docs/target/manpages/ .
+fi
 
 ## Copy over the puppet scripts
-tar -czf $WORKSPACE/deploy/target/puppet-scripts.tar.gz -C $WORKSPACE/deploy/packaging/puppet geowave
+if [ ! -f $WORKSPACE/deploy/target/puppet-scripts.tar.gz ]; then
+    tar -czf $WORKSPACE/deploy/target/puppet-scripts.tar.gz -C $WORKSPACE/deploy/packaging/puppet geowave
+fi
